Make 'svnadmnin recover' bump the repository instance ID because the admin might have copied the repo before attempting to recover it.
Under normal circumstances we would tell the admin to either use
hotcopy or to run 'svnadmin setuuid' on it.  The first might not
work on a corrupted repo while the second would invalidate working
copies even if the corruption did not affect the revision contents.
So, be nice to that stressed person and bump the instance ID.

* subversion/libsvn_fs_fs/recovery.c
  (recover_body): Bump instance ID before doing anything else.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1623381 13f79535-47bb-0310-9956-ffa450edef68