Continue fixing the 'svnfsfs stats' scalability issue with large f6 revs.
Read noderevs from the file individually. The complication here is that
we want to know the length of the noderev on disk as well.  A normal parser
call wouldn't give us that.  Therefore, we read the noderev into a stringbuf
and proceed with it just like we do for f7.

* subversion/libsvn_fs_fs/stats.c
  (read_noderev): We parse the NODEREV_STR only.
  (read_phsy_noderev): New utility function to read the noderev string
                       from the file.
  (parse_dir): We must read the noderev here as read_noderev no longer does
               it for us.
  (read_noderev): Simplify based on the knowledge that we get exactly the
                  noderev data passed in here.
  (read_phys_revision): Read the root noderev here instead of read_noderev().
  (read_log_rev_or_packfile): Update caller.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1639348 13f79535-47bb-0310-9956-ffa450edef68