Fix the FSFS-specific txn root cleanup misbehavior demonstrated by r1643138. Neither BDB nor FSX use txn-local caches.
The solution implemented here is simple: Both, FS->POOL and the pool owning
the txn, must be limit the lifetime for the TXN_DIR_CACHE to be valid.  If
one of them gets cleaned up, the cache needs to be removed as well.  So,
register one cleanup function per pool and, when one of the gets triggered,
unregister the respective other one.

* subversion/libsvn_fs_fs/caching.c
  (txn_cleanup_baton_t): Remember the pools that we registered cleanup on.
  (remove_txn_cache): Duplicate and rename to ...
  (remove_txn_cache_fs,
   remove_txn_cache_txn): ... these. Unregister the respective other cleanup.
  (init_txn_callbacks): Add FS parameter to have access to its pool.
  (svn_fs_fs__initialize_txn_caches): Update caller.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1643139 13f79535-47bb-0310-9956-ffa450edef68