Merge the pin-externals branch to trunk.
The discussion thread start in this post:
http://mail-archives.apache.org/mod_mbox/subversion-dev/201501.mbox/%3C20150128095434.GM26869%40ted.stsp.name%3E
has come to a consensus that the branch can be merged and additional
development of this feature can happen on trunk.

This merge adds a '--pin-externals' option to 'svn copy'.
This option enables automated pinning of URLs in svn:externals properties
during copy operations (issue #1258).

  --pin-externals          : pin externals with no explicit revision to their
                             last-changed revision (recommended when tagging)

This feature makes the svncopy.pl contrib script unnecessary.

Externals are "pinned" by adding a peg revision to the external's source URL.
For example, the external definition:
  ^/foo/bar ext_bar
might become:
  ^/foo/bar@400 ext_bar

This table shows which revision is used for pinning:

    copy source: working copy (WC)       REPOS
   ------------+------------------------+---------------------------+
    copy    WC | external's WC BASE rev | external's repos HEAD rev |
    dest:      |------------------------+---------------------------+
         REPOS | external's WC BASE rev | external's repos HEAD rev |
   ------------+------------------------+---------------------------+

An external that is already pinned is left as-is.

Please see the branch log for details about these changes.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1659395 13f79535-47bb-0310-9956-ffa450edef68