Add a new test showing a regression in commit over ra_serf where it fails to delete a directory tree containing many locked files.
The client-side error is: Unexpected HTTP status 400 'Bad Request' on '<path>'
The server logs this line: request failed: error reading the headers 

The problem is reproducible with 1.8.10 but not with 1.8.9.
It was introduced on the 1.8.x branch in r1606976.

In this test, the commit only fails with a certain number of locked files in
the WC. 50 files doesn't have the problem, 60 triggers the problem for me.
Currently, the test is using 75 for good measure.

Further investigation is pending. Help from ra_serf experts is appreciated :)

* subversion/tests/cmdline/lock_tests.py
  (delete_dir_with_lots_of_locked_files): New test which XFails over DAV and
   passes over all other RA layers.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1659536 13f79535-47bb-0310-9956-ffa450edef68