Following up on a list of recent blame changes, make the revisions in the 'kidney blame' make a bit more sense.
The revisions we show in blame are the revisions in which something
changed... but the revision we get in the callback is the revision of
which we have the actual file data.

To make things match up the way you would expect we have to delay reporting
the revisions until we receive the next delta.

This patch reverts r1661202, and implements that behavior.

* subversion/libsvn_client/blame.c
  (file_rev_baton): Add boolean and revision state.
  (file_rev_handler): Remove calculating the highest revision.
    Store the expected information in the revision baton, by
    delaying everything a bit. Reinstate old assertion.
  (svn_client_blame5): Remove ignore-revision code. Initialize baton.

* subversion/tests/cmdline/blame_tests.py
  (blame_youngest_to_oldest,
   blame_reverse_no_change): Update expected results.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1661208 13f79535-47bb-0310-9956-ffa450edef68