Fix a bad assertion in the diff callback wrappers that tried to check that a node wasn't added and existed at the same time.
This function will be called for adds, simple-changes and deletes, and
as such determines what happens based on these arguments. Because these
old diff callbacks didn't pass all information on their users the converted
drivers sometimes just forgot to pass information (such as copyfrom). One
of the drivers was recently fixed on 1.8.x and uncovered this problem.

* subversion/libsvn_wc/diff_editor.c
  (wrap_dir_opened): Fix bad assertion.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1667471 13f79535-47bb-0310-9956-ffa450edef68