Implement initial support for 'svn diff --summarize --ignore-properties'.
This is part of issue #4567.

We filter property changes, and paths which only have property changes, in the
svn client during diff display. This is not ideal because we want this feature
to be available in the API and we want to avoid adding further special cases
for different diff modes.
However, this simple patch makes a potential 1.9.x backport easier.
An implementation in libsvn_client is planned for later on trunk (1.10-dev)
and will supersede this patch.

See http://colabti.org/irclogger/irclogger_log/svn-dev?date=2015-03-23#l35
for IRC discussion.

Suggested by: julianfoad
              Chris <devnullaccount@yahoo.se>
http://mail-archives.apache.org/mod_mbox/subversion-dev/201503.mbox/%3C1427103800.22413.YahooMailBasic%40web121602.mail.ne1.yahoo.com%3E

* subversion/svn/diff-cmd.c
  (summarize_baton_t): Add ignore_properties flag.
  (summarize_xml): Filter paths with no text/tree (i.e. property-only) changes.
   Always show property modification status as "none".
  (summarize_regular): Filter paths with no text/tree (i.e. property-only) changes.
   Always show a blank property modification column.
  (svn_cl__diff): Set ignore_properties flag in summarize baton.

* subversion/tests/cmdline/diff_tests.py
  (diff_summarize_ignore_properties, test_list): New test.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1668625 13f79535-47bb-0310-9956-ffa450edef68