In FSFS, fix the delta combining code for the case that the delta chain starts with a PLAIN rep and some DELTA rep in the chain has 0 source ops, i.e. is "all new contents".
From 1.7 and earlier, we carried over an optimization that stops walking
the delta chain when a delta window would use any string from the older
delta window.  If the delta chain starts with an PLAIN rep, the new 1.8
code assumed that not having read a base rep implies that we need to read
said PLAIN rep.  However, in case of the short-cut kicking in, it would
wrongly read the omitted DELTA rep as PLAIN rep of fulltext window size.
The result would not be used (SCR_OPS=0) but the read may be beyond EOF,
triggering an I/O error.

So, this fixes a data re-construction bug, not a repo corruption.

* subversion/libsvn_fs_fs/cached_data.c
  (get_combined_window): Don't attempt to read the source window if we
                         already decided that we don't need it.

Reported by: rschupp

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1676667 13f79535-47bb-0310-9956-ffa450edef68