For FSFS, make 'svnadmin pack' lock the repo only if there is a shard to pack.
This is more a refactoring job than an actual scalability improvement.
As long as a pack is running, any additional pack requests will pile up
waiting for the first one to complete.  Only when that one is done, no
additional ones will queue while the already waiting ones are being
served as no-ops.

* subversion/libsvn_fs_fs/pack.c
  (get_pack_status): Logic factored out of pack_body.
  (pack_body): Call new function, remove pre-condition check here.
  (svn_fs_fs__pack): Check preconditions before acquiring the lock.
                     Also have an early out if there is nothing to do.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1686542 13f79535-47bb-0310-9956-ffa450edef68